requirements.txt
tensorflow>=2.0.0
numpy>=1.19.0
Pillow>=8.0.0
moviepy>=1.0.3
pathlib2>=2.3.0
